// src/cityjson/schemas/cityjson.ts
export interface CityJSONV201 {
// ...
  vertices: [number, number, number][];
  transform: {
    /**
     * @minItems 3
     * @maxItems 3
     */
    scale: [number, number, number];
    /**
     * @minItems 3
     * @maxItems 3
     */
    translate: [number, number, number];
  };
  appearance?: {
// ...
    "vertices-texture"?: [number, number][];
  };
// ...
}